Now only multisampling if samples > 1. Backbuffer now has no depth buffer anymore. Added "TextureObject" which is the super of RenderTexture/DepthTexture/Texture, so it can be used in shaders. Added RenderTexture/DepthTexture for managing these independent of framebuffers. Fixed draw; which should not use indices unless told so. Added subpixel blending (individual). Primitive buffer now matches layout if it has the same bindings with same formats, but missing formats are ignored (so you can render without primitive buffer for example). Removed the idea of "Surface"; Swapchain is a different beast than a Framebuffer. Theoretically added support for layered depth and render textures

+++ api/opengl/include/graphics/gl_graphics.hpp
@@ -51,7 +51,7 @@ namespace ignis {
 
 		//OpenGL constants
 
-		u32 maxSamples;
+		u8 maxSamples;
 		f32 maxAnistropy;
 
 		u32 major, minor;
